Use `concurrently` so `sends` is interrupted by Close, and add fs2 termination test
The previous fix moved `unNoneTerminate` from `receives` to the merged stream so
it'd terminate when the server's Close frame arrived even if the input stream
was still emitting. That solved the hang, but it also let fs2's `merge`
prefetch one extra `ws.receive()` before the halt propagated - which raises
`WebSocketClosed` because the httpclient backend arms a poison pill on the
queue right after returning a Close.

Switch to `receives.concurrently(sends.drain)` and keep `unNoneTerminate` on
`receives`. `concurrently` interrupts the `sends` background fiber as soon as
`receives` ends, and `unNoneTerminate` halts `receives` before any follow-up
`ws.receive()` can land on the poison pill.

Also rename `WebSocketSttpClientTests` to `WebSocketSttpClientFs2Tests` (the
Zio/Pekko siblings already follow this convention) and add a real-server test
that drives the pipe with `Stream("test") ++ Stream.never[IO]` against the
`ws/echo/fragmented` endpoint - asserting that the result stream terminates
with the expected reply even though the input is still active.
